* genget.c (LOWER): Cast argument to isupper()/tolower() to int
git-svn-id: svn://anonsvn.mit.edu/krb5/trunk@13380 dc483132-0cff-0310-8789-dd5450dbe970

+++ src/appl/telnet/libtelnet/genget.c
@@ -35,7 +35,7 @@
 
 #include <ctype.h>
 
-#define	LOWER(x) (isupper(x) ? tolower(x) : (x))
+#define	LOWER(x) (isupper((int) x) ? tolower((int) x) : (x))
 /*
  * The prefix function returns 0 if *s1 is not a prefix
  * of *s2.  If *s1 exactly matches *s2, the negative of
